+++ src/docs/user.yaml
@@ -66,67 +66,7 @@ paths:
                                 $ref: '#/components/schemas/ValidationError'
                 500:
                     description: Internal server error
-    /api/organization/:
-        put:
-            tags:
-                - Users
-            summery: Update an organization profile
-            description: Update information of an organization
-            operationId: updateOrga
-            consumes:
-                - multipart/form-data
-            produces:
-                - application/json
-            security:
-                - auth_cookie: []
-            requestBody:
-                required: true
-                content:
-                    multipart/form-data:
-                        schema:
-                            $ref: '#/components/schemas/UpdatedOrganization'
-                        encoding:
-                            categories:
-                                style: form
-                                explode: true
-            responses:
-                204:
-                    description: The updated organization.
-                    content:
-                        application/json:
-                            schema:
-                                $ref: '#/components/schemas/Organization'
-                403:
-                    description: User is not authorized to perform the operation.
-                    content:
-                        application/json:
-                            schema:
-                                type: object
-                                properties:
-                                    message:
-                                        type: string
-                                        example: User Not Authorised!
-                401:
-                    description: organization is not signed in.
-                    content:
-                        application/json:
-                            schema:
-                                type: object
-                                properties:
-                                    error:
-                                        type: boolean
-                                        example: true
-                                    message:
-                                        type: string
-                                        example: 'Invalid Token: No authorization token was found'
-                422:
-                    description: This status is returned when at least one of the optional fields has an invalid value.
-                    content:
-                        application/json:
-                            schema:
-                                $ref: '#/components/schemas/ValidationError'
-                500:
-                    description: Internal server error
+
 components:
     schemas:
         UpdatedUser:
@@ -269,140 +209,6 @@ components:
                             - Youth
                             - Justice
                     example: [No Poverty, Zero Hunger]
-        UpdatedOrganization:
-            type: object
-            properties:
-                email:
-                    type: string
-                    format: email
-                    example: test@gmail.com
-                password:
-                    type: string
-                    format: password
-                    example: 12341234
-                description:
-                    type: string
-                    example: This is a test organization
-                name:
-                    type: string
-                    example: Test Organization
-                city:
-                    type: string
-                    enum:
-                        - Adana
-                        - Kocaeli
-                        - Adiyaman
-                        - Konya
-                        - Afyonkarahisar
-                        - Kutahya
-                        - Agri
-                        - Malatya
-                        - Amasya
-                        - Manisa
-                        - Ankara
-                        - Kahramanmaras
-                        - Antalya
-                        - Mardin
-                        - Artviin
-                        - Mugla
-                        - Aydin
-                        - Mus
-                        - Balikesir
-                        - Nevsehir
-                        - Bilecik
-                        - Nigde
-                        - Bingol
-                        - Ordu
-                        - Bitlis
-                        - Rize
-                        - Bolu
-                        - Sakarya
-                        - Burdur
-                        - Samsun
-                        - Bursa
-                        - Siirt
-                        - Canakkale
-                        - Sinop
-                        - Cankiri
-                        - Sivas
-                        - Corum
-                        - Tekirdag
-                        - Denizli
-                        - Tokat
-                        - Diyarbakir
-                        - Trabzon
-                        - Edirne
-                        - Tunceli
-                        - Elazig
-                        - Sanliurfa
-                        - Erzincan
-                        - Usak
-                        - Erzurum
-                        - Van
-                        - Eskisehir
-                        - Yozgat
-                        - Gaziantep
-                        - Zonguldak
-                        - Giresun
-                        - Aksaray
-                        - Gumushane
-                        - Bayburt
-                        - Hakkari
-                        - Karaman
-                        - Hatay
-                        - Kirikkale
-                        - Isparta
-                        - Batman
-                        - Mersin
-                        - Sirnak
-                        - Istanbul
-                        - Bartin
-                        - Izmir
-                        - Ardahan
-                        - Kars
-                        - Igdir
-                        - Kastamonu
-                        - Yalova
-                        - Kayseri
-                        - Karabuk
-                        - Kirklareli
-                        - Kilis
-                        - Kirsehir
-                        - Osmaniye
-                        - Duzce
-                    example: Adana
-                categories:
-                    type: array
-                    items:
-                        type: string
-                        enum:
-                            - No Poverty
-                            - Zero Hunger
-                            - Good Health And Well-Being
-                            - Quality Education
-                            - Gender Equality
-                            - Clean Water And Sanitation
-                            - Affordable And Clean Energy
-                            - Animals
-                            - Oceans
-                            - Nature
-                            - Reduced Inequalities
-                            - Sustainable Cities And Communities
-                            - Responsible Consumption And Production
-                            - Climate Action
-                            - Life Below Water
-                            - Life On Land
-                            - Peace
-                            - Youth
-                            - Justice
-                        example: [No Poverty, Zero Hunger]
-                websiteUrl:
-                    type: string
-                    format: url
-                    example: https://www.test.com
-                coverImage:
-                    type: string
-                    format: binary
 
 securitySchemes:
     auth_cookie:
